+++ rpc/rpccore/subscription_cache.go
@@ -1,4 +1,4 @@
-package rpcv9
+package rpccore
 
 import "sync"
